backend_db/requirements.txt
Django==4.2.12
djangorestframework==3.15.2
paho-mqtt==2.1.0
django-filter==24.3
Markdown==3.7